feat: cs:app problem solving 2.59
