--- cs:app/ed/chapter2/src/hello-world.c
@@ -1,6 +1,0 @@
-#include <stdio.h>
-
-int main() {
-  printf("Hello, World! \n");
-  return 0;
-}
